Added game version
Planning on moving border stuff to PlanetScript and creating a list of GameObjects in that script which contains every GameObject that is in orbit of that specific planet.

The point of this is so we can have multiple planets the player (and maybe enemies as well) can move around on.