Changed some stuff
Changed the killBorderDistance of stage 1 from 40 to 120.
Changed the jump direction with joystick from up goes down and down goes up to up goes up and down goes down.

Jump joystick bug fixed
Changed the logic on which planet is the mainPlanetObj (planet with biggest G)
Added arrow for showing where Ducky is going to jump
Added animations (Projectiles of player and enemies, also the animation for the coin)
Removed the Universal Render Pipeline